use push distances as heuristic, accept test changes that are better (pushes/boxxle1/6 is worse)

+++ src/data.rs
@@ -60,9 +60,9 @@ impl Pos {
         Pos { r, c }
     }
 
-    crate fn dist(self, other: Pos) -> i16 {
-        (i16::from(self.r) - i16::from(other.r)).abs()
-            + (i16::from(self.c) - i16::from(other.c)).abs()
+    crate fn dist(self, other: Pos) -> u16 {
+        ((i16::from(self.r) - i16::from(other.r)).abs()
+            + (i16::from(self.c) - i16::from(other.c)).abs()) as u16
     }
 
     crate fn neighbors(self) -> [Pos; 4] {
